Move native application/library rule to base plugin
